+++ docs/source/index.rst
@@ -24,7 +24,7 @@ Welcome to Funsor's documentation!
    :caption: Interfaces:
 
    distributions
-   pyro_distributions
+   pyro
    minipyro
    einsum
