+++ funsor/pyro/__init__.py
@@ -1,0 +1,9 @@
+from funsor.pyro.distribution import FunsorDistribution
+from funsor.pyro.hmm import DiscreteDiscreteHMM, DiscreteHMM, GaussianGaussianHMM
+
+__all__ = [
+    "DiscreteDiscreteHMM",
+    "DiscreteHMM",
+    "FunsorDistribution",
+    "GaussianGaussianHMM",
+]
